refactor(BE): 소셜 인증 모듈 통합 및 구조 개선
