+++ back/controllers/SocialAuth/index.js
@@ -1,0 +1,5 @@
+import GoogleAuth from "./GoogleAuth.js";
+import NaverAuth from "./NaverAuth.js";
+import KakaoAuth from "./KakaoAuth.js";
+
+export { GoogleAuth, NaverAuth, KakaoAuth };
